+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased changes
 
+- Add timeout to downloading out of band catchup files. The timeout is controlled by the option `--download-blocks-timeout` (environment variable `CONCORDIUM_NODE_CONSENSUS_DOWNLOAD_BLOCKS_TIMEOUT`) and defaults to 5min.
+  block indices and catch-up chunk files specified by an URL. The default
+  timeout is 5 now minutes per chunk instead of waiting indefinitely.
 - Remove the "instrumentation" feature of the node and build the node with
   Prometheus support enabled by default.
   - Remove the `CONCORDIUM_NODE_PROMETHEUS_SERVER` environment variable.