+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Unreleased changes
 
 - Avoid deadlocks during node shutdown in specific scenarios.
+- The node will now shut down to start if an error occurs in a required service
+  (e.g., grpc server). In particular, the node will shut down if a required
+  service could not be started.
+- Add timeout to downloading out of band catchup files when block indices and
+  catch-up chunk files are specified by an URL. The timeout is controlled
+  by the option `--download-blocks-timeout` (environment variable
+  `CONCORDIUM_NODE_CONSENSUS_DOWNLOAD_BLOCKS_TIMEOUT`) and defaults to 5 min.
+  timeout is 5 now minutes per chunk instead of waiting indefinitely.
 - Remove the "instrumentation" feature of the node and build the node with
   Prometheus support enabled by default.
   - Remove the `CONCORDIUM_NODE_PROMETHEUS_SERVER` environment variable.
@@ -11,7 +19,7 @@
 
 ## 5.1.1
 
-- Relay blocks earlier. In particular this means that blocks are now processed in 
+- Relay blocks earlier. In particular this means that blocks are now processed in
   two steps, `block receive` and `block execute`. The former performs verification of block meta data
   while the latter adds the block to the tree.
   Blocks are now enqueued in the outgoing message queue in between the the two steps.